+++ pywaffle/__init__.py
@@ -2,7 +2,8 @@
 # -*-coding: utf-8 -*-
 
 from ._version import __version__
+from .fontawesome_handler import font_awesome_status
 from .functional import waffle_chart
 from .waffle import Waffle
 
-__all__ = ["Waffle", "waffle_chart", "__version__"]
+__all__ = ["Waffle", "waffle_chart", "font_awesome_status", "__version__"]